Disallow occurrence searches with >100k results
We've encountered an issue where MySQL can fill up /var/tmp writing filesort output when handling overly broad queries.
This is tricky to solve in general since there are not going to be good indexes for every possible search. We can revisit
if this becomes an issue for legitimate users or if we decide on some ways to constrain the advanced search