Fix password reset link
In 9e4909ad00b0b7cc1a0f3a4b0e96ddb2839c0067, the id of the login input changed, which broke the Javascript that pops up an informational alert when the login is empty. This makes it impossible to know how password reset is supposed to work, since it's not clear that it uses the form above.